requirements.txt
Flask==2.2.5
requests==2.31.0
yfinance==0.2.20
pandas==1.5.3
nltk==3.8.1
openai==0.27.8
gunicorn==20.1.0
APScheduler==3.9.1
python-dotenv==1.0.0
# Optional dependencies below
# Comment out if installation issues occur
# finbert-embedding>=0.1.0
# transformers>=4.28.0
# torch>=2.0.0 